Implemented C porting of the FFT class from WACUP (See link in comments) and created first visualization module rendered via Direct2D (To be completed)

+++ WinAudio_Wasapi/WA_CircleBuffer.h
@@ -9,7 +9,7 @@ typedef struct tagWA_CircleBuffer WA_CircleBuffer;
 /// </summary>
 /// <param name="uBufferSize">Size in bytes of the buffer</param>
 /// <returns>NULL on fail or a Pointer to a WA_CircleBuffer Instance</returns>
-WA_CircleBuffer* WA_CircleBuffer_New(uint16_t uBufferSize);
+WA_CircleBuffer* WA_CircleBuffer_New(uint32_t uBufferSize);
 
 /// <summary>
 /// Free memory allocated with WA_CircleBuffer_New function
@@ -24,7 +24,7 @@ void WA_CircleBuffer_Delete(WA_CircleBuffer* This);
 /// <param name="pBuffer">Pointer to a Byte Buffer</param>
 /// <param name="uBufferSize">Size of the Buffer</param>
 /// <returns>True on Success</returns>
-bool WA_CircleBuffer_Write(WA_CircleBuffer* This, int8_t* pBuffer, uint16_t uBufferSize);
+bool WA_CircleBuffer_Write(WA_CircleBuffer* This, int8_t* pBuffer, uint32_t uBufferSize);
 
 /// <summary>
 /// Read Data from the Circle Buffer
@@ -34,7 +34,7 @@ bool WA_CircleBuffer_Write(WA_CircleBuffer* This, int8_t* pBuffer, uint16_t uBuf
 /// <param name="uBufferSize">Size of the Buffer</param>
 /// <param name="uPosition">Position in bytes from which to start reading </param>
 /// <returns>True on Success</returns>
-bool WA_CircleBuffer_ReadFrom(WA_CircleBuffer* This, int8_t* pBuffer, uint16_t uBufferSize, uint16_t uPosition);
+bool WA_CircleBuffer_ReadFrom(WA_CircleBuffer* This, int8_t* pBuffer, uint32_t uBufferSize, uint32_t uPosition);
 
 /// <summary>
 /// Reset Write Index to Position 0
